05_Our_First_Agent_with_LangGraph/pyproject.toml
[project]
name = "05-our-first-agent-with-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "fastapi>=0.118.0",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "langsmith>=0.4.30",
    "openevals>=0.1.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.4",
    "tavily-python>=0.7.12",
    "uvicorn>=0.37.0",
]
